+++ pysweep/core/solver.py
@@ -157,7 +157,7 @@ class Solver(object):
         """Use this function to begin the simulation."""
         # -------------------------------SWEPT RULE---------------------------------------------#
         #setting global time step to zero
-        self.globalTimeStep=numpy.int32(1) #Has to be int32 for GPU
+        self.globalTimeStep=1 #Has to be int32 for GPU
         # -------------------------------FIRST PRISM AND COMMUNICATION-------------------------------------------#
         functions.FirstPrism(self)
         functions.firstForward(self)
@@ -178,7 +178,7 @@ class Solver(object):
     def standardSolve(self):
         # -------------------------------Standard Decomposition---------------------------------------------#
         #setting global time step to zero
-        self.globalTimeStep=numpy.int32(1)
+        self.globalTimeStep=1
         #Send Boundary points
         functions.sendEdges(self)
         cwt = 0 #Starts at zero compared too swept because of the write algorithm
